+++ docs/Menu bar.md
@@ -1,0 +1,27 @@
+# Menu Bar
+
+## Public
+
+Shown when user is not logged in.
+
+* {Logo} Title (`~/`)
+* Sign In (`~/Account/Login`)
+
+## Staff
+
+Shown when staff is logged in.
+
+* {Logo} (`~/Staff`)
+* New Complaint (`~/Staff/Complaints/Add`)
+* Complaint Search (`~/Staff/Complaints`)
+* Action Search (`~/Staff/ComplaintActions`)
+* More (Drop-down)
+    * Reports (`~/Admin/Reports`)
+    * CTS Users (`~/Admin/Users`)
+    * Site Maintenance (`~/Admin/Maintenance`)
+    * Public Portal (`~/Public`)
+* Account (Drop-down)
+    * You profile (`~/Account`)
+    * Support (`~/Account/Support`)
+    * Sign out (*form*)
+* Theme toggle
