+++ docs/Redirects.md
@@ -1,41 +1,39 @@
 # Site URL Redirects
 
-This document defines routes from the old version of the application that should be redirected according to the
-new [Site Map](Site%20map.md).
+This document defines incomplete URLs or routes from the old version of the application that should be redirected
+according to the [Site Map](Site%20map.md).
 
 ## Public pages
 
-| Redirect                        | To                        |
-|---------------------------------|---------------------------|
-| `/Public/ComplaintDetails/{id}` | `/Public/Complaints/{id}` |
+| Redirect                            | To                                     |
+|-------------------------------------|----------------------------------------|
+| `/Public`                           | `/Index`                               |
+| `/Public/ComplaintDetails/{id}`     | `/Complaint/{id}`                      |
+| `/Public/Attachment/{attachmentId}` | `/Complaint/Attachment/{attachmentId}` |
+| `/Complaint`                        | `/Index`                               |
 
-## Complaint pages
+## Staff complaint pages
 
-| Redirect                             | To                       |
-|--------------------------------------|--------------------------|
-| `/Complaints`                        | `/Staff/Complaints`      |
-| `/Complaints/Details/{id}`           | `/Staff/Complaints/{id}` |
-| `/Complaints/[workflow action]/{id}` | `/Staff/Complaints/{id}` |
-| `/Staff/Complaints/Details`          | `/Staff/Complaints`      |
+| Redirect                                | To                                 |
+|-----------------------------------------|------------------------------------|
+| `/Complaints`                           | `/Staff/Index`                     |
+| `/Complaints/Details/{id}`              | `/Staff/Complaints/{id}`           |
+| `/Complaints/Attachment/{attachmentId}` | `/Staff/Attachment/{attachmentId}` |
+| `/Staff/Complaints`                     | `/Staff/Index`                     |
 
 ## Complaint Actions
 
-| Redirect                                | To                        |
-|-----------------------------------------|---------------------------|
-| `/ComplaintActions`                     | `/Staff/ComplaintActions` |
-| `/Complaints/Actions/{id}`              | `/Staff/ComplaintActions` |
-| `/Complaints/EditAction/{actionid}`     | `/Staff/ComplaintActions` |
-| `/Complaints/Details/{id}/DeleteAction` | `/Staff/Complaints/{id}`  |
-| `/Complaints/Details/{id}/EditAction`   | `/Staff/Complaints/{id}`  |
+| Redirect                   | To                               |
+|----------------------------|----------------------------------|
+| `/ComplaintActions`        | `/Staff/ComplaintActions/Index`  |
+| `/Complaints/Actions/{id}` | `/Staff/Complaints/{id}#actions` |
 
 ## Users
 
-| Redirect               | To                          |
-|------------------------|-----------------------------|
-| `/Users`               | `/Admin/Users`              |
-| `/Users/Details/{id}`  | `/Admin/Users/Details/{id}` |
-| `/Users/Edit/{id}`     | `/Admin/Users/Details/{id}` |
-| `/Admin/Users/Details` | `/Admin/Users`              |
+| Redirect              | To                          |
+|-----------------------|-----------------------------|
+| `/Users`              | `/Admin/Users/Index`        |
+| `/Users/Details/{id}` | `/Admin/Users/Details/{id}` |
 
 ## Site Maintenance
 
@@ -43,9 +41,9 @@ new [Site Map](Site%20map.md).
 |-----------------------------------|-------------------------------------------|
 | `/Maintenance`                    | `/Admin/Maintenance`                      |
 | `/[maintenance option]`           | `/Admin/Maintenance/[maintenance option]` |
-| `/[maintenance option]/Create`    | `/Admin/Maintenance/[maintenance option]` |
 | `/[maintenance option]/Details/*` | `/Admin/Maintenance/[maintenance option]` |
-| `/[maintenance option]/Edit/*`    | `/Admin/Maintenance/[maintenance option]` |
-| `/Reports`                        | `/Admin/Reports`                          |
-| `/Reports/[report type]`          | `/Admin/Reports/[report type]`            |
-| `/Export`                         | `/Admin/Reports/Export`                   |
+| `/Reports`                        | `/Admin/Reporting` *                      |
+| `/Reports/[report type]`          | `/Admin/Reporting/[report type]` *        |
+| `/Export`                         | `/Admin/Reporting/Export`                 |
+
+* If possible, given implementation of ArcGIS.
