+++ src/AppServices/Staff/StaffService.cs
@@ -13,6 +13,7 @@ namespace Cts.AppServices.Staff;
 public sealed class StaffService(
     IUserService userService,
     UserManager<ApplicationUser> userManager,
+    // ReSharper disable once SuggestBaseTypeForParameterInConstructor
     IMapper mapper,
     IOfficeRepository officeRepository)
     : IStaffService
