+++ src/WebApp/Pages/Account/Login.cshtml
@@ -7,7 +7,7 @@
 <h1>@ViewData["Title"]</h1>
 <p>
     The Complaint Tracking System administration site is a State of Georgia application.
-    It is intended for conducting official State business and is available for use by
+    It is provided for conducting official State business and is available for use by
     State of Georgia employees only.
 </p>
