+++ src/WebApp/Pages/Attachment.cshtml
@@ -1,0 +1,2 @@
+﻿@page "~/Complaint/Attachment/{id:Guid}/{fileName=''}"
+@model AttachmentModel
