--- src/WebApp/Pages/Public/Complaints/Attachment.cshtml
@@ -1,2 +1,0 @@
-﻿@page "{id:Guid}/{fileName=''}"
-@model AttachmentModel
