+++ src/WebApp/Pages/Shared/Components/MainMenu/Default.cshtml
@@ -26,7 +26,7 @@
                     else
                     {
                         <li class="nav-item">
-                            <a class="nav-link" asp-page="/Public/Index">Search</a>
+                            <a class="nav-link" asp-page="/Index">Search</a>
                         </li>
                     }
                 </ul>
