+++ src/WebApp/Pages/Staff/ComplaintActions/Edit.cshtml
@@ -40,7 +40,7 @@
 
         <div class="mt-3 mb-1">
             <button type="submit" class="btn btn-primary col-6 col-sm-4 col-lg-3 me-2">Save Changes</button>
-            <a asp-page="Details" asp-route-id="@Model.ComplaintView.Id.ToString()" asp-fragment="@Model.ActionItemId" class="btn btn-outline-secondary">Cancel</a>
+            <a asp-page="../Complaints/Details" asp-route-id="@Model.ComplaintView.Id.ToString()" asp-fragment="@Model.ActionItemId" class="btn btn-outline-secondary">Cancel</a>
         </div>
         <div class="text-danger-emphasis">* denotes a required field</div>
         <input asp-for="ActionItemId" type="hidden" />
