+++ src/WebApp/Pages/Staff/ComplaintActions/Restore.cshtml
@@ -36,7 +36,7 @@
     <form method="post">
         <div class="mt-3 mb-1">
             <button type="submit" class="btn btn-danger col-5 col-sm-3 col-lg-2 me-2">Restore</button>
-            <a asp-page="Details" asp-route-id="@Model.ActionItemView.ComplaintId.ToString()" asp-fragment="actions" class="btn btn-outline-secondary">Cancel</a>
+            <a asp-page="../Complaints/Details" asp-route-id="@Model.ActionItemView.ComplaintId.ToString()" asp-fragment="@Model.ActionItemId.ToString()" class="btn btn-outline-secondary">Cancel</a>
         </div>
         <input asp-for="ActionItemId" type="hidden" />
     </form>
