+++ src/WebApp/Pages/Staff/Complaints/Attachment/Delete.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "{attachmentId:guid?}"
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-@model DeleteAttachmentModel
+@model AttachmentDeleteModel
 
 @{
     ViewData["Title"] = $"Delete Attachment from Complaint ID {Model.ComplaintId}";
